rename app module
